refactor based on offline discussion with eric
