+++ StitchCore-iOS/StitchCore-iOS/Core/Internal/StitchAppClientImpl.swift
@@ -61,7 +61,7 @@ internal final class StitchAppClientImpl: StitchAppClient {
             try StitchAuthImpl.init(
                 requestClient: StitchRequestClientImpl.init(baseURL: config.baseURL,
                                                             transport: config.transport,
-                                                            transportTimeout: config.transportTimeout),
+                                                            defaultRequestTimeout: config.defaultRequestTimeout),
                 authRoutes: self.routes.authRoutes,
                 storage: config.storage,
                 dispatcher: self.dispatcher,
@@ -131,6 +131,33 @@ internal final class StitchAppClientImpl: StitchAppClient {
                              _ completionHandler: @escaping (Any?, Error?) -> Void) {
         dispatcher.run(withCompletionHandler: completionHandler) {
             return try self.coreClient.callFunctionInternal(withName: name, withArgs: args)
+        }
+    }
+
+    /**
+     * Calls the MongoDB Stitch function with the provided name and arguments, as well as with a specified timeout. Use
+     * this for functions that may run longer than the default 15 second timeout.
+     *
+     * - parameters:
+     *     - withName: The name of the Stitch function to be called.
+     *     - withArgs: The `BSONArray` of arguments to be provided to the function.
+     *     - withRequestTimeout: The number of seconds the client should wait for a response from the server before
+     *                           failing with an error.
+     *     - completionHandler: The completion handler to call when the function call is complete.
+     *                          This handler is executed on a non-main global `DispatchQueue`.
+     *     - result: The result of the function call as an `Any`, or `nil` if the function call failed.
+     *     - error: An error object that indicates why the function call failed, or `nil` if the function call was
+     *              successful.
+     *
+     */
+    public func callFunction(withName name: String,
+                             withArgs args: BSONArray,
+                             withRequestTimeout requestTimeout: TimeInterval,
+                             _ completionHandler: @escaping (Any?, Error?) -> Void) {
+        dispatcher.run(withCompletionHandler: completionHandler) {
+            return try self.coreClient.callFunctionInternal(withName: name,
+                                                            withArgs: args,
+                                                            withRequestTimeout: requestTimeout)
         }
     }
 }
