+++ StitchCore-iOS/StitchCore-iOS/Core/StitchAppClient.swift
@@ -1,4 +1,5 @@
 import ExtendedJSON
+import StitchCore
 
 /**
  * The fundamental set of methods for communicating with a MongoDB Stitch application.
@@ -61,6 +62,24 @@ public protocol StitchAppClient {
      *
      */
     func callFunction(withName name: String, withArgs args: BSONArray, _ completionHandler: @escaping (_ result: Any?, _ error: Error?) -> Void)
+
+    /**
+     * Calls the MongoDB Stitch function with the provided name and arguments, as well as with a specified timeout. Use
+     * this for functions that may run longer than the default 15 second timeout.
+     *
+     * - parameters:
+     *     - withName: The name of the Stitch function to be called.
+     *     - withArgs: The `BSONArray` of arguments to be provided to the function.
+     *     - withRequestTimeout: The number of seconds the client should wait for a response from the server before
+     *                           failing with an error.
+     *     - completionHandler: The completion handler to call when the function call is complete.
+     *                          This handler is executed on a non-main global `DispatchQueue`.
+     *     - result: The result of the function call as an `Any`, or `nil` if the function call failed.
+     *     - error: An error object that indicates why the function call failed, or `nil` if the function call was
+     *              successful.
+     *
+     */
+    func callFunction(withName name: String, withArgs args: BSONArray, withRequestTimeout requestTimeout: TimeInterval, _ completionHandler: @escaping (_ result: Any?, _ error: Error?) -> Void)
 
     // swiftlint:enable line_length
 }
