+++ StitchCore/Sources/StitchCore/StitchAppClientConfiguration.swift
@@ -51,13 +51,13 @@ public struct StitchAppClientConfigurationImpl: StitchAppClientConfiguration, Bu
     public let transport: Transport
     
     /**
-     * The number of seconds that a `Transport` should spend on an HTTP round trip before failing with an error.
+     * The number of seconds that a `Transport` should spend by default on an HTTP round trip before failing with an
+     * error.
      *
-     * - important: If the underlying transport internally handles timeouts, it is possible that the transport will
-     *              throw a timeout error before this specified interval. If you experience premature timeouts,
-     *              configure your underlying transport to have a longer timeout interval.
+     * - important: If a request timeout was specified for a specific operation, for example in a function call, that
+     *              timeout will override this one.
      */
-    public let transportTimeout: TimeInterval
+    public let defaultRequestTimeout: TimeInterval
 
     /**
      * The client app id of the Stitch application that this client is going to communicate with.
@@ -115,15 +115,15 @@ public struct StitchAppClientConfigurationImpl: StitchAppClientConfiguration, Bu
             throw StitchClientConfigurationError.missingTransport
         }
         
-        guard let transportTimeout = builder.transportTimeout else {
-            throw StitchClientConfigurationError.missingTransportTimeout
+        guard let defaultRequestTimeout = builder.defaultRequestTimeout else {
+            throw StitchClientConfigurationError.missingDefaultRequestTimeout
         }
 
         self.baseURL = baseURL
         self.dataDirectory = dataDirectory
         self.storage = storage
         self.transport = transport
-        self.transportTimeout = transportTimeout
+        self.defaultRequestTimeout = defaultRequestTimeout
     }
 }
 
@@ -165,13 +165,13 @@ public struct StitchAppClientConfigurationBuilder: StitchClientConfigurationBuil
     public var transport: Transport?
     
     /**
-     * The number of seconds that a `Transport` should spend on an HTTP round trip before failing with an error.
+     * The number of seconds that a `Transport` should spend by default on an HTTP round trip before failing with an
+     * error.
      *
-     * - important: If the underlying transport internally handles timeouts, it is possible that the transport will
-     *              throw a timeout error before this specified interval. If you experience premature timeouts,
-     *              configure your underlying transport to have a longer timeout interval.
+     * - important: If a request timeout was specified for a specific operation, for example in a function call, that
+     *              timeout will override this one.
      */
-    public var transportTimeout: TimeInterval?
+    public var defaultRequestTimeout: TimeInterval?
 
     /**
      * The client app id of the Stitch application that this client is going to communicate with.
