+++ StitchCore/Sources/StitchCore/StitchError.swift
@@ -22,10 +22,9 @@ public enum StitchError: Error {
      * transport errors, these errors are thrown by the underlying `Transport` of the Stitch client, and thus contain
      * the error that the transport threw. Errors in decoding the result from the server include the specific error
      * thrown when attempting to decode the response. An error code is included, which indicates whether the error
-     * was a transport error or decoding error. If the underlying transport times out based on the configured
-     * timeout interval, there will be no underying error specified.
+     * was a transport error or decoding error.
      */
-    case requestError(withError: Error?, withRequestErrorCode: StitchRequestErrorCode)
+    case requestError(withError: Error, withRequestErrorCode: StitchRequestErrorCode)
 
     /**
      * Indicates that an error occurred when using the Stitch client, typically before the client performed a request.
@@ -94,7 +93,6 @@ public enum StitchServiceErrorCode: String, Codable {
  */
 public enum StitchRequestErrorCode {
     case transportError
-    case transportTimeoutError
     case decodingError
 }
