+++ StitchCore/Tests/stitch-ios-sdk-protoTests/StitchAppClientConfigurationTests.swift
@@ -9,7 +9,7 @@ class StitchAppClientConfigurationBuilderTests: XCTestCase {
     private let dataDirectory = URL.init(string: "foo/bar/baz/qux")!
     private let storage = MemoryStorage.init()
     private let transport = FoundationHTTPTransport.init()
-    private let transportTimeout: TimeInterval = defaultTestTransportTimeout
+    private let defaultRequestTimeout: TimeInterval = testDefaultRequestTimeout
 
     func testStitchAppClientConfigurationBuilderInit() throws {
         var builder = StitchAppClientConfigurationBuilder { _ in }
@@ -53,10 +53,10 @@ class StitchAppClientConfigurationBuilderTests: XCTestCase {
         
         XCTAssertThrowsError(try builder.build()) { error in
             XCTAssertEqual(error as? StitchClientConfigurationError,
-                           StitchClientConfigurationError.missingTransportTimeout)
+                           StitchClientConfigurationError.missingDefaultRequestTimeout)
         }
         
-        builder.transportTimeout = self.transportTimeout
+        builder.defaultRequestTimeout = self.defaultRequestTimeout
 
         let config = try builder.build()
 
@@ -66,6 +66,6 @@ class StitchAppClientConfigurationBuilderTests: XCTestCase {
         XCTAssertEqual(config.baseURL, self.baseURL)
         XCTAssert(config.storage is MemoryStorage)
         XCTAssert(config.transport is FoundationHTTPTransport)
-        XCTAssertEqual(config.transportTimeout, self.transportTimeout)
+        XCTAssertEqual(config.defaultRequestTimeout, self.defaultRequestTimeout)
     }
 }
