+++ StitchCore/Tests/stitch-ios-sdk-protoTests/stitch_ios_sdk_protoTests.swift
@@ -62,7 +62,7 @@ public final class AtomicPort {
 
 let atomicPort = AtomicPort()
 
-let defaultTestTransportTimeout: TimeInterval = 15.0
+let testDefaultRequestTimeout: TimeInterval = 15.0
 
 private func start(server: HttpServer) -> UInt16 {
     repeat {
